ms_tree_hub/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1;python_version<"3.9"
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-cors-headers==3.14.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.1.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
openapi-codec==1.3.2
packaging==23.0
psycopg2-binary==2.9.5
pytz==2023.2
PyYAML==6.0.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.4
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.15
zipp==3.15.0
